Fixed todos
Also some minor refactorings.

Some changes might be breaking.
