+++ api/api.go
@@ -6,6 +6,7 @@ package api
 
 import (
 	"context"
+	impl "github.com/spiffe/spike-sdk-go/api/internal/impl/api"
 
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
 
@@ -13,10 +14,12 @@ import (
 	"github.com/spiffe/spike-sdk-go/spiffe"
 )
 
+// Api is the SPIKE API.
 type Api struct {
 	source *workloadapi.X509Source
 }
 
+// New creates a new SPIKE API.
 func New() *Api {
 	defaultEndpointSocket := spiffe.EndpointSocket()
 
@@ -28,68 +31,85 @@ func New() *Api {
 	return &Api{source: source}
 }
 
+// NewWithSource creates a new SPIKE API with a given X509Source.
 func NewWithSource(source *workloadapi.X509Source) *Api {
 	return &Api{source: source}
 }
 
+// Close closes the SPIKE API.
 func (a *Api) Close() {
 	spiffe.CloseSource(a.source)
 }
 
+// CreatePolicy creates a new policy.
 func (a *Api) CreatePolicy(
 	name string, spiffeIdPattern string, pathPattern string,
 	permissions []data.PolicyPermission,
 ) error {
-	return CreatePolicy(a.source,
+	return impl.CreatePolicy(a.source,
 		name, spiffeIdPattern, pathPattern, permissions)
 }
 
+// DeletePolicy deletes a policy.
 func (a *Api) DeletePolicy(name string) error {
-	return DeletePolicy(a.source, name)
+	return impl.DeletePolicy(a.source, name)
 }
 
+// GetPolicy gets a policy.
 func (a *Api) GetPolicy(name string) (*data.Policy, error) {
-	return GetPolicy(a.source, name)
+	return impl.GetPolicy(a.source, name)
 }
 
+// ListPolicies lists all policies.
 func (a *Api) ListPolicies() (*[]data.Policy, error) {
-	return ListPolicies(a.source)
+	return impl.ListPolicies(a.source)
 }
 
+// DeleteSecretVersions deletes secret versions.
 func (a *Api) DeleteSecretVersions(path string, versions []int) error {
-	return DeleteSecret(a.source, path, versions)
+	return impl.DeleteSecret(a.source, path, versions)
 }
 
+// DeleteSecret deletes a secret.
 func (a *Api) DeleteSecret(path string) error {
-	return DeleteSecret(a.source, path, []int{})
+	return impl.DeleteSecret(a.source, path, []int{})
 }
 
-func (a *Api) GetSecretVersioned(path string, version int) (*data.Secret, error) {
-	return GetSecret(a.source, path, version)
+// GetSecretVersioned gets a secret versioned.
+func (a *Api) GetSecretVersioned(
+	path string, version int,
+) (*data.Secret, error) {
+	return impl.GetSecret(a.source, path, version)
 }
 
+// GetSecret gets a secret.
 func (a *Api) GetSecret(path string) (*data.Secret, error) {
-	return GetSecret(a.source, path, 0)
+	return impl.GetSecret(a.source, path, 0)
 }
 
+// ListSecretKeys lists all secret key paths
 func (a *Api) ListSecretKeys() (*[]string, error) {
-	return ListSecretKeys(a.source)
+	return impl.ListSecretKeys(a.source)
 }
 
+// GetSecretMetadata gets secret metadata.
 func (a *Api) GetSecretMetadata(
 	path string, version int,
 ) (*data.SecretMetadata, error) {
-	return GetSecretMetadata(a.source, path, version)
+	return impl.GetSecretMetadata(a.source, path, version)
 }
 
+// PutSecret puts a secret.
 func (a *Api) PutSecret(path string, data map[string]string) error {
-	return PutSecret(a.source, path, data)
+	return impl.PutSecret(a.source, path, data)
 }
 
+// UndeleteSecret undeletes secret versions.
 func (a *Api) UndeleteSecret(path string, versions []int) error {
-	return UndeleteSecret(a.source, path, versions)
+	return impl.UndeleteSecret(a.source, path, versions)
 }
 
+// Init initializes SPIKE Nexus.
 func (a *Api) Init() error {
-	return Init(a.source)
+	return impl.Init(a.source)
 }
