+++ api/internal/impl/api/acl_create.go
@@ -1,8 +1,8 @@
-package api
-
 //    \\ SPIKE: Secure your secrets with SPIFFE.
 //  \\\\\ Copyright 2024-present SPIKE contributors.
 // \\\\\\\ SPDX-License-Identifier: Apache-2.0
+
+package api
 
 import (
 	"encoding/json"
@@ -10,42 +10,48 @@ import (
 
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
 
+	"github.com/spiffe/spike-sdk-go/api/entity/data"
 	"github.com/spiffe/spike-sdk-go/api/entity/v1/reqres"
 	"github.com/spiffe/spike-sdk-go/api/internal/url"
 	"github.com/spiffe/spike-sdk-go/net"
 )
 
-func DeletePolicy(source *workloadapi.X509Source, id string) error {
-	r := reqres.PolicyDeleteRequest{
-		Id: id,
+func CreatePolicy(source *workloadapi.X509Source,
+	name string, spiffeIdPattern string, pathPattern string,
+	permissions []data.PolicyPermission,
+) error {
+	r := reqres.PolicyCreateRequest{
+		Name:            name,
+		SpiffeIdPattern: spiffeIdPattern,
+		PathPattern:     pathPattern,
+		Permissions:     permissions,
 	}
 
 	mr, err := json.Marshal(r)
 	if err != nil {
 		return errors.Join(
 			errors.New(
-				"deletePolicy: I am having problem generating the payload",
+				"createPolicy: I am having problem generating the payload",
 			),
 			err,
 		)
 	}
 
-	var truer = func(string) bool { return true }
-	client, err := net.CreateMtlsClient(source, truer)
+	client, err := net.CreateMtlsClient(source)
 	if err != nil {
 		return err
 	}
 
-	body, err := net.Post(client, url.PolicyDelete(), mr)
+	body, err := net.Post(client, url.PolicyCreate(), mr)
 	if err != nil {
 		return err
 	}
 
-	res := reqres.PolicyDeleteResponse{}
+	res := reqres.PolicyCreateResponse{}
 	err = json.Unmarshal(body, &res)
 	if err != nil {
 		return errors.Join(
-			errors.New("deletePolicy: Problem parsing response body"),
+			errors.New("createPolicy: Problem parsing response body"),
 			err,
 		)
 	}
