+++ api/internal/impl/api/acl_delete.go
@@ -10,49 +10,41 @@ import (
 
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
 
-	"github.com/spiffe/spike-sdk-go/api/entity/data"
 	"github.com/spiffe/spike-sdk-go/api/entity/v1/reqres"
 	"github.com/spiffe/spike-sdk-go/api/internal/url"
 	"github.com/spiffe/spike-sdk-go/net"
 )
 
-func CreatePolicy(source *workloadapi.X509Source,
-	name string, spiffeIdPattern string, pathPattern string,
-	permissions []data.PolicyPermission,
-) error {
-	r := reqres.PolicyCreateRequest{
-		Name:            name,
-		SpiffeIdPattern: spiffeIdPattern,
-		PathPattern:     pathPattern,
-		Permissions:     permissions,
+func DeletePolicy(source *workloadapi.X509Source, id string) error {
+	r := reqres.PolicyDeleteRequest{
+		Id: id,
 	}
 
 	mr, err := json.Marshal(r)
 	if err != nil {
 		return errors.Join(
 			errors.New(
-				"createPolicy: I am having problem generating the payload",
+				"deletePolicy: I am having problem generating the payload",
 			),
 			err,
 		)
 	}
 
-	var truer = func(string) bool { return true }
-	client, err := net.CreateMtlsClient(source, truer)
+	client, err := net.CreateMtlsClient(source)
 	if err != nil {
 		return err
 	}
 
-	body, err := net.Post(client, url.PolicyCreate(), mr)
+	body, err := net.Post(client, url.PolicyDelete(), mr)
 	if err != nil {
 		return err
 	}
 
-	res := reqres.PolicyCreateResponse{}
+	res := reqres.PolicyDeleteResponse{}
 	err = json.Unmarshal(body, &res)
 	if err != nil {
 		return errors.Join(
-			errors.New("createPolicy: Problem parsing response body"),
+			errors.New("deletePolicy: Problem parsing response body"),
 			err,
 		)
 	}
