+++ api/internal/impl/api/acl_get.go
@@ -27,8 +27,7 @@ func GetPolicy(source *workloadapi.X509Source, id string) (*data.Policy, error) 
 		)
 	}
 
-	var truer = func(string) bool { return true }
-	client, err := net.CreateMtlsClient(source, truer)
+	client, err := net.CreateMtlsClient(source)
 	if err != nil {
 		return nil, err
 	}
