+++ api/internal/impl/api/secret_undelete.go
@@ -30,7 +30,7 @@ import (
 //
 // Example:
 //
-//	err := UndeleteSecret(x509Source, "secret/path", []string{"1", "2"})
+//	err := undeleteSecret(x509Source, "secret/path", []string{"1", "2"})
 func UndeleteSecret(source *workloadapi.X509Source,
 	path string, versions []int) error {
 	var vv []int
@@ -53,8 +53,7 @@ func UndeleteSecret(source *workloadapi.X509Source,
 		)
 	}
 
-	var truer = func(string) bool { return true }
-	client, err := net.CreateMtlsClient(source, truer)
+	client, err := net.CreateMtlsClient(source)
 	if err != nil {
 		return err
 	}
